+++ apps/functions-subscription/src/domain/__tests__/mocks.ts
@@ -1,29 +1,27 @@
-import { vi } from 'vitest';
+import { mock, mockFn } from 'vitest-mock-extended';
+import { SubscriptionReader, SubscriptionWriter } from '../subscription';
+import { SubscriptionRequestWriter } from '../subscription-request';
+import {
+  SubscriptionHistoryReader,
+  SubscriptionHistoryWriter,
+} from '../subscription-history';
+import { EventWriter } from '../event';
+import { ActivationRequestRepository } from '../activation-request';
+import { Clock } from '../clock';
+import { MonotonicIdFn } from '../monotonic-id';
+import { HashFn } from '../hash';
 
 export const makeTestEnv = () => {
-  const subscriptionReaderMock = {
-    get: vi.fn(),
-  };
-  const subscriptionWriterMock = {
-    insert: vi.fn(),
-  };
-  const subscriptionRequestWriterMock = {
-    insert: vi.fn(),
-  };
-  const subscriptionHistoryWriterMock = {
-    insert: vi.fn(),
-  };
-  const hashFnMock = vi.fn();
-  const clockMock = {
-    now: vi.fn(),
-  };
-
   return {
-    subscriptionReader: subscriptionReaderMock,
-    subscriptionWriter: subscriptionWriterMock,
-    subscriptionRequestWriter: subscriptionRequestWriterMock,
-    subscriptionHistoryWriter: subscriptionHistoryWriterMock,
-    hashFn: hashFnMock,
-    clock: clockMock,
+    subscriptionReader: mock<SubscriptionReader>(),
+    subscriptionWriter: mock<SubscriptionWriter>(),
+    subscriptionRequestWriter: mock<SubscriptionRequestWriter>(),
+    subscriptionHistoryReader: mock<SubscriptionHistoryReader>(),
+    subscriptionHistoryWriter: mock<SubscriptionHistoryWriter>(),
+    activationRequestRepository: mock<ActivationRequestRepository>(),
+    eventWriter: mock<EventWriter>(),
+    clock: mock<Clock>(),
+    hashFn: mockFn<HashFn>(),
+    monotonicIdFn: mockFn<MonotonicIdFn>(),
   };
 };