+++ apps/functions-subscription/src/system-env.ts
@@ -3,6 +3,8 @@ import { Capabilities } from './domain/capabilities';
 import { insertSubscription } from './use-cases/insert-subscription';
 import { getSubscription } from './use-cases/get-subscription';
 import { processSubscriptionRequest } from './use-cases/process-subscription-request';
+import { processActivationJob } from './use-cases/process-activation-job';
+import { processActivationRequest } from './use-cases/process-activation-request';
 
 export const makeSystemEnv = (capabilities: Capabilities) => ({
   insertSubscription: flow(insertSubscription, apply(capabilities)),
@@ -11,6 +13,8 @@ export const makeSystemEnv = (capabilities: Capabilities) => ({
     processSubscriptionRequest,
     apply(capabilities),
   ),
+  processActivationJob: flow(processActivationJob, apply(capabilities)),
+  processActivationRequest: flow(processActivationRequest, apply(capabilities)),
 });
 
 export type SystemEnv = ReturnType<typeof makeSystemEnv>;