+++ infra/resources/modules/commons/function_async.tf
@@ -16,13 +16,28 @@ locals {
     FETCH_KEEPALIVE_FREE_SOCKET_TIMEOUT = "30000"
     FETCH_KEEPALIVE_TIMEOUT             = "60000"
 
-    COSMOSDB_ENDPOINT                  = module.cosmosdb_account.endpoint
-    COSMOSDB_DATABASE_NAME             = module.cosmosdb_sql_database_trial.name
-    EVENTHUB_NAMESPACE                 = "${module.event_hub.name}.servicebus.windows.net"
-    SUBSCRIPTION_REQUEST_EVENTHUB_NAME = "${local.domain}-subscription-requests"
-    SUBSCRIPTION_REQUEST_CONSUMER      = "on"
+    COSMOSDB_ENDPOINT      = module.cosmosdb_account.endpoint
+    COSMOSDB_DATABASE_NAME = module.cosmosdb_sql_database_trial.name
+    EVENTHUB_NAMESPACE     = "${module.event_hub.name}.servicebus.windows.net"
+    SERVICEBUS_NAMESPACE   = "${local.servicebus_namespace}.servicebus.windows.net"
 
+    LEASES_COSMOSDB_CONTAINER_NAME = azurerm_cosmosdb_sql_container.leases.name
+
+    SUBSCRIPTION_HISTORY_CONSUMER                        = "on"
+    SUBSCRIPTION_HISTORY_COSMOSDB_CONTAINER_NAME         = azurerm_cosmosdb_sql_container.subscription_history.name
+    SubscriptionHistoryCosmosConnection__accountEndpoint = module.cosmosdb_account.endpoint
+
+    SUBSCRIPTION_REQUEST_CONSUMER                                  = "on"
+    SUBSCRIPTION_REQUEST_EVENTHUB_NAME                             = "${local.domain}-subscription-requests"
     SubscriptionRequestEventHubConnection__fullyQualifiedNamespace = "${module.event_hub.name}.servicebus.windows.net"
+
+    ACTIVATION_CONSUMER                                   = "on"
+    ACTIVATION_MAX_FETCH_SIZE                             = "999"
+    ACTIVATIONS_COSMOSDB_CONTAINER_NAME                   = azurerm_cosmosdb_sql_container.activations.name
+    ActivationConsumerCosmosDBConnection__accountEndpoint = module.cosmosdb_account.endpoint
+
+    EVENTS_PRODUCER              = "on"
+    EVENTS_SERVICEBUS_TOPIC_NAME = "${local.domain}-topic-events"
   }
 }
 
@@ -82,7 +97,11 @@ module "subscription_async_fn" {
 
   app_settings = merge(
     local.async_app_settings,
-    {},
+    {
+      # Avoiding host ID collisions
+      # https://learn.microsoft.com/en-us/azure/azure-functions/storage-considerations?tabs=azure-cli#avoiding-host-id-collisions
+      AzureFunctionsWebHost__hostid = "subscription-async-fn-01"
+    },
   )
 
   sticky_app_setting_names = []
@@ -123,6 +142,13 @@ module "subscription_async_fn" {
   ]
 
   tags = var.tags
+}
+
+# Enables the subscription_async_fn to write on service-bus topic
+resource "azurerm_role_assignment" "subs_asyn_write_to_sbt" {
+  scope                = azurerm_servicebus_topic.events.id
+  role_definition_name = "Azure Service Bus Data Sender"
+  principal_id         = module.subscription_async_fn.system_identity_principal
 }
 
 # Enables the subscription_async_fn to read from the event-hub
@@ -174,6 +200,13 @@ module "subscription_async_fn_staging_slot" {
   system_identity_enabled = true
 
   tags = var.tags
+}
+
+# Enables the subscription_async_fn_staging to write on service-bus topic
+resource "azurerm_role_assignment" "subs_asyn_staging_write_to_sbt" {
+  scope                = azurerm_servicebus_topic.events.id
+  role_definition_name = "Azure Service Bus Data Sender"
+  principal_id         = module.subscription_async_fn_staging_slot.system_identity_principal
 }
 
 # Enables the subscription_async_fn_staging to read from the event-hub