+++ infra/resources/modules/commons/function_subscription.tf
@@ -16,11 +16,25 @@ locals {
     FETCH_KEEPALIVE_FREE_SOCKET_TIMEOUT = "30000"
     FETCH_KEEPALIVE_TIMEOUT             = "60000"
 
-    COSMOSDB_ENDPOINT                  = module.cosmosdb_account.endpoint
-    COSMOSDB_DATABASE_NAME             = module.cosmosdb_sql_database_trial.name
-    EVENTHUB_NAMESPACE                 = "${module.event_hub.name}.servicebus.windows.net"
-    SUBSCRIPTION_REQUEST_EVENTHUB_NAME = "${local.domain}-subscription-requests"
+    COSMOSDB_ENDPOINT      = module.cosmosdb_account.endpoint
+    COSMOSDB_DATABASE_NAME = module.cosmosdb_sql_database_trial.name
+    EVENTHUB_NAMESPACE     = "${module.event_hub.name}.servicebus.windows.net"
+    SERVICEBUS_NAMESPACE   = "${local.servicebus_namespace}.servicebus.windows.net"
+
+    LEASES_COSMOSDB_CONTAINER_NAME = azurerm_cosmosdb_sql_container.leases.name
+
+    SUBSCRIPTION_HISTORY_CONSUMER                = "off"
+    SUBSCRIPTION_HISTORY_COSMOSDB_CONTAINER_NAME = azurerm_cosmosdb_sql_container.subscription_history.name
+
     SUBSCRIPTION_REQUEST_CONSUMER      = "off"
+    SUBSCRIPTION_REQUEST_EVENTHUB_NAME = "${local.domain}-subscription-requests"
+
+    ACTIVATION_CONSUMER                 = "off"
+    ACTIVATION_MAX_FETCH_SIZE           = "999"
+    ACTIVATIONS_COSMOSDB_CONTAINER_NAME = azurerm_cosmosdb_sql_container.activations.name
+
+    EVENTS_PRODUCER              = "off"
+    EVENTS_SERVICEBUS_TOPIC_NAME = "${local.domain}-topic-events"
   }
 }